requirements.txt
tkinter
Pillow
transformers 